fix: swap slop-native and slop-mcp-proxy naming
The names were backwards:
- slop-native = per-affordance dynamic tools (native tool experience)
- slop-mcp-proxy = generic app_action proxy tool
