+++ packages/typescript/integrations/claude/slop-mcp-proxy/servers/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "slop-bridge",
+  "name": "slop-bridge-mcp-proxy",
   "version": "0.1.0",
   "private": true,
   "type": "module",
